Security enhancements
Added rights check in every resource
Refactored AuthUser class

+++ bpm-console-rest/src/main/java/pl/edu/pw/ii/bpmConsole/rest/resources/deployments/DeploymentResource.java
@@ -4,8 +4,9 @@ import io.dropwizard.auth.Auth;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import pl.edu.pw.ii.bpmConsole.interfaces.DeploymentService;
-import pl.edu.pw.ii.bpmConsole.rest.filters.BpmUser;
 import pl.edu.pw.ii.bpmConsole.rest.resources.LinkBuilder;
+import pl.edu.pw.ii.bpmConsole.rest.resources.RightsVerification;
+import pl.edu.pw.ii.bpmConsole.valueObjects.AuthUser;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -15,7 +16,7 @@ import javax.ws.rs.core.Response;
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 @Path("/deployments/{processDefinitionId}")
-public class DeploymentResource {
+public class DeploymentResource implements RightsVerification {
 
     DeploymentService deploymentService;
 
@@ -25,7 +26,8 @@ public class DeploymentResource {
     }
 
     @DELETE
-    public Response delete(@Auth BpmUser user, @PathParam("processDefinitionId") String processDefinitionId) {
+    public Response delete(@Auth AuthUser user, @PathParam("processDefinitionId") String processDefinitionId) {
+        verifyAdminRights(user);
         deploymentService.delete(processDefinitionId);
         return Response
                 .noContent()
