+++ bpm-console-rest/src/main/java/pl/edu/pw/ii/bpmConsole/rest/resources/user/GroupsResource.java
@@ -4,8 +4,9 @@ import io.dropwizard.auth.Auth;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import pl.edu.pw.ii.bpmConsole.interfaces.UserService;
-import pl.edu.pw.ii.bpmConsole.rest.filters.BpmUser;
 import pl.edu.pw.ii.bpmConsole.rest.resources.LinkBuilder;
+import pl.edu.pw.ii.bpmConsole.rest.resources.RightsVerification;
+import pl.edu.pw.ii.bpmConsole.valueObjects.AuthUser;
 import pl.edu.pw.ii.bpmConsole.valueObjects.GroupInfo;
 
 import javax.validation.Valid;
@@ -17,57 +18,62 @@ import javax.ws.rs.core.Response;
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 @Path("/groups")
-public class GroupResource {
+public class GroupsResource implements RightsVerification {
     private final UserService userService;
 
     @Autowired
-    public GroupResource(UserService userService) {
+    public GroupsResource(UserService userService) {
         this.userService = userService;
     }
 
     @GET
-    public Response list(@Auth BpmUser user) {
+    public Response list(@Auth AuthUser user) {
+        verifyAdminRights(user);
         return Response
                 .ok(userService.listGroups())
-                .links(LinkBuilder.fromResource(GroupResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(GroupsResource.class).rel("self").build())
                 .build();
     }
 
     @GET
     @Path("/{id}")
-    public Response get(@Auth BpmUser user, @PathParam("id") String groupId) {
+    public Response get(@Auth AuthUser user, @PathParam("id") String groupId) {
+        verifyAdminRights(user);
         return Response
                 .ok(userService.getGroup(groupId))
-                .links(LinkBuilder.fromResource(GroupResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(GroupsResource.class).rel("self").build())
                 .build();
     }
 
     @POST
-    public Response create(@Auth BpmUser user, @Valid GroupInfo groupInfo) {
+    public Response create(@Auth AuthUser user, @Valid GroupInfo groupInfo) {
+        verifyAdminRights(user);
         userService.createGroup(groupInfo);
         return Response
                 .ok()
-                .links(LinkBuilder.fromResource(GroupResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(GroupsResource.class).rel("self").build())
                 .build();
     }
 
     @PUT
     @Path("/{id}")
-    public Response edit(@Auth BpmUser user, @PathParam("id") String groupId, @Valid GroupInfo groupInfo) {
+    public Response edit(@Auth AuthUser user, @PathParam("id") String groupId, @Valid GroupInfo groupInfo) {
+        verifyAdminRights(user);
         userService.editGroup(groupInfo, groupId);
         return Response
                 .ok()
-                .links(LinkBuilder.fromResource(GroupResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(GroupsResource.class).rel("self").build())
                 .build();
     }
 
     @DELETE
     @Path("/{id}")
-    public Response delete(@Auth BpmUser user, @PathParam("id") String groupId) {
+    public Response delete(@Auth AuthUser user, @PathParam("id") String groupId) {
+        verifyAdminRights(user);
         userService.deleteGroup(groupId);
         return Response
                 .ok()
-                .links(LinkBuilder.fromResource(GroupResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(GroupsResource.class).rel("self").build())
                 .build();
     }
 }
