+++ bpm-console-rest/src/main/java/pl/edu/pw/ii/bpmConsole/rest/resources/user/UsersResource.java
@@ -4,8 +4,9 @@ import io.dropwizard.auth.Auth;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import pl.edu.pw.ii.bpmConsole.interfaces.UserService;
-import pl.edu.pw.ii.bpmConsole.rest.filters.BpmUser;
 import pl.edu.pw.ii.bpmConsole.rest.resources.LinkBuilder;
+import pl.edu.pw.ii.bpmConsole.rest.resources.RightsVerification;
+import pl.edu.pw.ii.bpmConsole.valueObjects.AuthUser;
 import pl.edu.pw.ii.bpmConsole.valueObjects.UserInfo;
 
 import javax.validation.Valid;
@@ -17,66 +18,72 @@ import javax.ws.rs.core.Response;
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 @Path("/users")
-public class UserResource {
+public class UsersResource implements RightsVerification {
     private final UserService userService;
 
     @Autowired
-    public UserResource(UserService userService) {
+    public UsersResource(UserService userService) {
         this.userService = userService;
     }
 
     @GET
-    public Response list(@Auth BpmUser user) {
+    public Response list(@Auth AuthUser user) {
+        verifyAdminRights(user);
         return Response
                 .ok(userService.listUsers())
-                .links(LinkBuilder.fromResource(UserResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(UsersResource.class).rel("self").build())
                 .build();
     }
 
     @GET
     @Path("/{id}")
-    public Response get(@Auth BpmUser user, @PathParam("id") String userId) {
+    public Response get(@Auth AuthUser user, @PathParam("id") String userId) {
+        userService.verifyRights(user).toUser(userId);
         return Response
                 .ok(userService.getUser(userId))
-                .links(LinkBuilder.fromResource(UserResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(UsersResource.class).rel("self").build())
                 .build();
     }
 
     @POST
-    public Response create(@Auth BpmUser user, @Valid UserInfo userInfo) {
+    public Response create(@Auth AuthUser user, @Valid UserInfo userInfo) {
+        verifyAdminRights(user);
         userService.createUser(userInfo);
         return Response
                 .ok()
-                .links(LinkBuilder.fromResource(UserResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(UsersResource.class).rel("self").build())
                 .build();
     }
 
     @PUT
     @Path("/{id}")
-    public Response edit(@Auth BpmUser user, @PathParam("id") String userId, @Valid UserInfo userInfo) {
+    public Response edit(@Auth AuthUser user, @PathParam("id") String userId, @Valid UserInfo userInfo) {
+        userService.verifyRights(user).toUser(userId);
         userService.editUser(userInfo, userId);
         return Response
                 .ok()
-                .links(LinkBuilder.fromResource(UserResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(UsersResource.class).rel("self").build())
                 .build();
     }
 
     @DELETE
     @Path("/{id}")
-    public Response delete(@Auth BpmUser user, @PathParam("id") String userId) {
+    public Response delete(@Auth AuthUser user, @PathParam("id") String userId) {
+        verifyAdminRights(user);
         userService.deleteUser(userId);
         return Response
                 .ok()
-                .links(LinkBuilder.fromResource(UserResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(UsersResource.class).rel("self").build())
                 .build();
     }
 
     @GET
     @Path("/{id}/groups")
-    public Response getUserGroups(@Auth BpmUser user, @PathParam("id") String userId) {
+    public Response getUserGroups(@Auth AuthUser user, @PathParam("id") String userId) {
+        userService.verifyRights(user).toUser(userId);
         return Response
                 .ok(userService.getUserGroups(userId))
-                .links(LinkBuilder.fromResource(GroupResource.class).rel("self").build())
+                .links(LinkBuilder.fromResource(UsersResource.class).rel("self").build())
                 .build();
     }
 }
